% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/times4.R
\name{times4}
\alias{times4}
\title{Title Times number 4}
\usage{
times4(x = 3)
}
\arguments{
\item{x}{A numeric value}
}
\value{
Output the input number multiplied by 4
}
\description{
Title Times number 4
}
\examples{
times4(3)

}
